[IMP] stock, mrp: relocate location constraint to relevant module
The `_check_default_location()` constraint exists to prevent the user
from moving freshly produced goods to the scrapyard, which only makes
sense if production moves exist. (ie `mrp` is installed)

Task ID: 4385357
